% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/addons.R
\name{actionvariety}
\alias{actionvariety}
\title{get data for the top N variety of actions across all action categories}
\usage{
actionvariety(veris, topn = 20)
}
\arguments{
\item{veris}{a verisr object}

\item{topn}{how many entries to return}
}
\description{
Given a verisr object, this will return the \code{topn} most frequent action
varieties across all of the action categories.
}

