% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verisr.R
\name{buildVeris}
\alias{buildVeris}
\title{recursively build a data structure to store for the veris record}
\usage{
buildVeris(columns, nrow)
}
\arguments{
\item{columns}{list. The output of veriscol listing columns & 
types.  Likely 'vft'}

\item{nrow}{integer. The number of rows to create (likely 'numfil'
or 'length(jfiles)')}
}
\value{
a list of data tables and lists to fill
}
\description{
This recursively builds a list of the form 
list(DT, name1=list(), name2=list(), etc) where name is the name
of a column which will be a list of data frames and list() is
the same recursive structure, starting with a DT.
}
