\name{countMatrix}
\alias{countMatrix}
\title{count incidents in clusters}
\usage{
countMatrix(veris, label, filter = NULL)
}
\arguments{
  \item{veris}{a verisr object}

  \item{label}{a list of labels to count}

  \item{filter}{an optional filter to apply}
}
\description{
given a verisr object and a list of labels, this will
create a data.frame of the data suitable for plotting.
}

