% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/addons.R
\name{discovery_pie}
\alias{discovery_pie}
\title{Special for Wade Baker, get a pie chart of the internal/external discovery methods}
\usage{
discovery_pie(veris)
}
\arguments{
\item{veris}{a verisr object}
}
\description{
Given a verisr object, this will return a ggplot object of a pie chart showing
the proportion of internal versus external discovery methods.
}

