% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getenumCI2021.R
\name{flatten_verisr}
\alias{flatten_verisr}
\title{flatten a verisr dataframe such that each row represents a single plus.master_id}
\usage{
flatten_verisr(veris)
}
\arguments{
\item{veris}{A verisr object}
}
\value{
a verisr object with each plus.master_id represented by a single row
}
\description{
flatten a verisr dataframe such that each row represents a single plus.master_id
}
