% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verisr.R
\name{geta4names}
\alias{geta4names}
\title{Convenience function for the a4 names and values}
\usage{
geta4names()
}
\description{
This returns a named vector where the names are the column names 
in the final verisr data table and the valus are suitable for using
in a regex in the existing column names.
}
