% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/addons.R
\name{getall_discovery}
\alias{getall_discovery}
\title{Get the data for discovery methods from a verisr object}
\usage{
getall_discovery(veris)
}
\arguments{
\item{veris}{a verisr object}
}
\description{
Given a verisr object, this will return the discovery methods and values
for "All External" and "All Internal" summaries
}

