\name{getccc}
\alias{getccc}
\title{Creates a matrix of measurements about various fields within a veris matrix object.
will return a matrix of the same number of rows, but the columns are various measurements
around the completeness, context and complexity (CCC) of the incidents.}
\usage{
getccc(mat)
}
\arguments{
  \item{mat}{a verisr oject, should have removed
  unknown/other fields}
}
\description{
Creates a matrix of measurements about various fields
within a veris matrix object. will return a matrix of the
same number of rows, but the columns are various
measurements around the completeness, context and
complexity (CCC) of the incidents.
}

