% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/addons.R
\name{getdatavariety}
\alias{getdatavariety}
\title{get variety of data compromised}
\usage{
getdatavariety(veris)
}
\arguments{
\item{veris}{a verisr object}
}
\description{
given a verisr object, this will return a ggplot2 object showing the
amount of records lost for various data types over time.
}

