% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getenumSRC.R, R/verisr.R
\name{getenumCI}
\alias{getenumCI}
\alias{getenumby}
\alias{getenum}
\title{The main method for summarizes veris enumerations from verisr objects}
\usage{
getenumCI(...)

getenumby(...)

getenum(...)
}
\arguments{
\item{...}{Arguments passed on to \code{getenumCI2019}
\describe{
  \item{veris}{A verisr object}
  \item{enum}{A veris feature or enumeration to summarize}
  \item{by}{A veris feature or enumeration to group by}
  \item{na.rm}{A boolean of whether to include not applicable in the sample set.
This is REQUIRED if enum has a potential value of NA as there is no 
'default' method for handling NAs.  Instead, it depends on the 
hypothesis being tested.}
  \item{unk}{A boolean referring whether to include 'unknown' in the sample.
The default is 'FALSE' and should rarely be overwritten.}
  \item{short.names}{A boolean identifying whether to use the full enumeration
name or just the last section. (i.e. action.hacking.variety.SQLi vs
just SQLi.)}
  \item{ci.method}{A confidence interval method to use.  Current supported
methods are any from binom.confint() or "multinomial".  If unsure
which to use, use "wilson".}
  \item{ci.level}{A number from 0 to 1 representing the width of the 
confidence interval. (default = 0.95)}
  \item{round.freq}{An integer indicating how many places to round
the frequency value to. (default = 5)}
  \item{na}{DEPRECIATED! Use 'na.rm' parameter.}
}}
}
\description{
The main method for summarizes veris enumerations from verisr objects
}
\section{Functions}{
\itemize{
\item \code{getenumby}: getenumby A wrapper for getenumCI().  Maintained for backwards compatibility.

\item \code{getenum}: getenum A wrapper for getenumCI().  Maintained for backwards compatibility.
}}

