% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/addons.R
\name{getindemp}
\alias{getindemp}
\alias{getindemp.final}
\title{Specific view of victim industry and employee count}
\usage{
getindemp(veris)
}
\arguments{
\item{veris}{a verisr object}
}
\description{
Given a verisr object, these functions will query for the top level (2-digit)
industry codes, convert to a string and correlate those with the employee count
enumerations for the victims.
}
\details{
*getindemp* will get the information and plot it using the \code{\link{showtile}} function.

*getindemp.final* will not plot the information, but will return a data.frame
of the data and it will compress the employee count down to Small, Large and Unknown
}

