% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getenum.R
\name{getlastaction}
\alias{getlastaction}
\title{Get the last element from a column name, include action}
\usage{
getlastaction(nm)
}
\arguments{
\item{nm}{the vector of column names}
}
\description{
Givn a vector with one or more column names (veris fields), this will
return the last string in the name, as it is seperated by [.].  It will
also assume this is the action and attempt to map nice names to the name.
}

