\name{getlevel}
\alias{getlevel}
\title{return a vector for ordered factors within VERIS.}
\usage{
getlevel(enum, unknown = T)
}
\arguments{
  \item{enum}{a verisr field name}

  \item{unknown}{logical; include unknown in the
  enumeration}
}
\description{
given a VERIS enumeration field, return a vector that can
be used to sort the factors of the enumerations.
}

