% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getlogical}
\alias{getlogical}
\title{Convenience function to pull the logical columns from verisr object}
\usage{
getlogical(veris)
}
\arguments{
\item{veris}{the verisr object}
}
\description{
Given a verisr object this will return a vector of column names which are logical values.
}

