% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verisr.R
\name{getnth}
\alias{getnth}
\title{Get the nth element from a column name}
\usage{
getnth(nm, which = 2)
}
\arguments{
\item{nm}{the vector of column names}

\item{which}{the nth vlue to return}
}
\description{
Givn a vector with one or more column names (veris fields), this will
return the nth string in the name, as it is seperated by [.].
}

