% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/addons.R
\name{gettimeline.dbir}
\alias{gettimeline.dbir}
\title{Gets the data for a "detection deficit" plot}
\usage{
gettimeline.dbir(veris)
}
\arguments{
\item{veris}{verisr object}
}
\description{
Given a verisr object, it will return a data.frame for the percentage of
compromise and detection events within days or less and over time.
}

