% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/addons.R
\name{gettimeline}
\alias{gettimeline}
\title{work with timeline data from veris}
\usage{
gettimeline(mat, timelines = c("compromise", "exfiltration", "discovery",
  "containment"), justdata = F)
}
\arguments{
\item{mat}{the verisr object}

\item{timelines}{vector of one or more of the following: c("compromise", "exfiltration", "discovery", "containment")}

\item{justdata}{boolean, if TRUE it will return a dataframe of the data, else a ggplot2 object}
}
\description{
Munges and returns either data or a ggplot object (depending on the \code{justdata} argument)
from one or more categories for timeline measurements.
}

