% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{getVarietyAmount}
\alias{getVarietyAmount}
\title{Internal: Expand variety}
\usage{
getVarietyAmount(x)
}
\arguments{
\item{x}{a slice of a veris a verisr object}
}
\description{
This will expand all of the "variety" fields by the amount
specified in the "amount" field of the same level object.
}
