% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/csv.R
\name{getvnames}
\alias{getvnames}
\title{Internal: Get fields names}
\usage{
getvnames(veris)
}
\arguments{
\item{veris}{a verisr object}
}
\description{
This will grab all the field names from a veris object
}

