\name{indname2label}
\alias{indname2label}
\title{convert veris industry field names to short labels}
\usage{
indname2label(industry)
}
\arguments{
  \item{industry}{a vector of 2 or 3 digit veris industry
  names}
}
\description{
given a vector of veris names with industries,
(victim.industry.44 for example), returns a character
vector of short labels as an ordered factor.
}

