% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verisr.R
\docType{data}
\name{industry2}
\alias{industry2}
\title{Metadata for 2-digit NAICS industry classification}
\description{
This data allows a mapping between two digit NAICS code and the 
full definition provided by the NAICS specification and a shorter
version of the title for compact visuals.
}
\references{
\url{www.census.gov/naics/}
}
\keyword{data}
