\name{industry2label}
\alias{industry2label}
\title{convert industry codes to short labels}
\usage{
industry2label(industry)
}
\arguments{
  \item{industry}{a vector of 2 or 3 digit verisr object}
}
\description{
given a vector of 2 or 3 digit industry codes, returns a
character vector of short labels as an ordered factor.
}

