% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verisr.R
\name{joinVeris}
\alias{joinVeris}
\title{Join the list of data tables from buildVeris into a data frame}
\usage{
joinVeris(veris)
}
\arguments{
\item{veris}{list of data tables in buildVeris structure}
}
\value{
data frame. Data frame with list columns
}
\description{
This effectively produces the recursive veris data structure
}

