% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAutils.R
\name{logit2prob}
\alias{logit2prob}
\title{Convert logit to probability}
\usage{
logit2prob(logit)
}
\arguments{
\item{logit}{a vector of logits}
}
\value{
a vector of probabilities
}
\description{
Convert logit to probability
}
