% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns2021.R
\name{models_to_centroids}
\alias{models_to_centroids}
\title{Helper function to retrieve prototypes from skmeans models for add_patterns}
\usage{
models_to_centroids()
}
\arguments{
\item{breach}{A skmeans model of breach centroids. (Should)}
}
\value{
A matrix of of skmeans centroids with one row per centroid
}
\description{
Helper function to retrieve prototypes from skmeans models for add_patterns
}
