% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verisr.R
\name{nameveris}
\alias{nameveris}
\title{Return a dense and complete list of VERIS variables and values}
\usage{
nameveris(json, a4, vtype)
}
\arguments{
\item{json}{the json object from reading in a file.}

\item{a4}{the a4 object from geta4names()}

\item{vtype}{the vtype object from parseProps()}
}
\description{
Given an incident in json format, it will process the file and 
return a simple list object where each element is named with 
the field name and the value is the value read in, or True if the
field is an enumeration.
}

