\docType{data}
\name{orgize}
\alias{orgize}
\title{Metadata for victim orgsize}
\description{
  This data allows a mapping between the enumeration of
  victim employee_count and the designation of "large" and
  "small"
}
\references{
  \url{www.veriscommunity.net}
}
\keyword{data}
\keyword{datasets}

