% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verisr.R
\name{parseProps}
\alias{parseProps}
\title{Map VERIS fields to data type.}
\usage{
parseProps(schema, cur = "", outvec = NULL)
}
\arguments{
\item{schema}{the merged veris schema in json}

\item{cur}{the current name (internal)}

\item{outvec}{the current output vector (internal)}
}
\description{
Given a json schema for VERIS, this function will return a named vector
where the name is the field and the value is the R mode string or "enum".
}
\keyword{json}
