% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns2021.R
\name{pattern_1.3.6_to_1.3.5}
\alias{pattern_1.3.6_to_1.3.5}
\title{Groom data for patterns from VERIS 1.3.6 to 1.3.5}
\usage{
pattern_1.3.6_to_1.3.5(veris, centroids)
}
\arguments{
\item{veris}{veris object to cluster}

\item{centroids}{skmeans cluster prototypes}
}
\value{
veris, centroids
}
\description{
Because VERIS changes, columns come and go.  This makes the clustering
function unhappy.  We need to make sure the data we're clustering matches
the matrix we're using to cluster (based on veris 1.3.5).
}
