% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns2021.R
\name{pattern_current_to_1.3.5}
\alias{pattern_current_to_1.3.5}
\title{Currently a reference to verisr::pattern_1.3.6_to_1.3.5()}
\usage{
pattern_current_to_1.3.5(...)
}
\description{
Currently a reference to verisr::pattern_1.3.6_to_1.3.5()
}
