% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns2021.R
\name{pattern_no_change}
\alias{pattern_no_change}
\title{Groom data for patterns, no changes}
\usage{
pattern_no_change(veris, centroids)
}
\arguments{
\item{veris}{veris object to cluster}

\item{centroids}{skmeans cluster prototypes}
}
\value{
veris, centroids
}
\description{
Because VERIS changes, columns come and go.  This makes the clustering
function unhappy.  This function can be used if no no changes to the 
centroids are needed but the function is desired
}
