\name{pcamatrix}
\alias{pcamatrix}
\title{fold rows of matrix for further analysis}
\usage{
pcamatrix(veris, filter = NULL)
}
\arguments{
  \item{veris}{a verisr object}

  \item{filter}{an optional filter to apply}
}
\description{
given a verisr object and a list of features, this will
create a matrix of the data suitable for further PCA
analysis.
}

