% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{placeholder}
\alias{placeholder}
\title{Subset a hierarchical verisr object to a flattened verisr object}
\usage{
placeholder(x)
}
\arguments{
\item{veris}{The verisr object to flatten}

\item{columns}{}

\item{level}{Used during recursion}

\item{row.name}{Column to indicate unique rows in parent dataframe}
}
\value{
flattened, subsetted verisr object
}
\description{
(A flattened verisr object has no list columns)  Additionally,
only columns specified in the columns parameter will be included
in the dataframe
}
