\name{plot.verisr}
\alias{plot.verisr}
\title{Displays a four panel barplot of a verisr object}
\usage{
  \method{plot}{verisr} (object, ...)
}
\arguments{
  \item{object}{veris object to summarise}

  \item{...}{other arguments ignored (for compatibility
  with generic)}
}
\description{
  Displays a four panel barplot of a verisr object
}
\keyword{internal}

