% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plota4}
\alias{plota4}
\title{Given a verisr object it will create a ggplot of the a4 grid}
\usage{
plota4(veris, title = "A4 Grid", freq = T)
}
\arguments{
\item{veris}{the verisr object}

\item{title}{the title in the plot}

\item{freq}{logical, display frequency on the plot}
}
\description{
This will create the a4 grid (see Verizon's Data Breach Inves 2012) and return a ggplot object.
}

