% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/addons.R
\name{plotjchunk}
\alias{plotjchunk}
\title{Plot the object returned from setjenum}
\usage{
plotjchunk(chunk, final = F)
}
\arguments{
\item{chunk}{the object returned by \code{\link{setjenum}}}

\item{final}{boolean, whether we want to make it pretty (TRUE) or detailed (FALSE)}
}
\description{
This will plot the data returned from a call to \code{\link{setjenum}}.
Internally, it will call \code{\link{simplebar}} or \code{\link{simplejbar}}
depending if \code{final} is TRUE or FALSE respectively.
}
\seealso{
setjenum
}

