% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{recurse_columns}
\alias{recurse_columns}
\title{Find the locations of vectors in a hierarchical veris object}
\usage{
recurse_columns(veris, columns, level = NULL)
}
\arguments{
\item{veris}{The verisr object to flatten}

\item{columns}{to locate}

\item{level}{Used during recursion}
}
\value{
flattened, subsetted verisr object
}
\description{
Find the locations of vectors in a hierarchical veris object
}
