% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/addons.R
\name{setjenum}
\alias{setjenum}
\title{Sets up an object of multiple enumerations}
\usage{
setjenum(veris, enums, trim = 10, unknowns = c("Unknown", " - Other"))
}
\arguments{
\item{veris}{the verisr data}

\item{enums}{one or more enumeration to query (e.g. c("victim.orgsize", "victim.industry2"))}

\item{trim}{maximum values to return for any single enumeration, set to NULL to not trim}

\item{unknowns}{vector of strings to treat as "Unknown" values to not be counted in "n"}
}
\value{
a list with the following objects:
  * *ucount* : count of unknown in enumeration
  * *df* : the data frame(s) returned by getenum()
  * *title* : vector of suitable titles for plotting
  * *enum* : the specific enum for this chunk
  * *height* : vector with the height of each plot
  * *ptall* : the max height the entire plot, useful for knitr calls
}
\description{
Given a verisr object and one or more enumerations to query, it will set up an
object that enables plotting by \code{\link{plotjchunk}}.
The object also contains information that will help set the optimal height
of the object (for example, in a knitr document) and information about the
data queried.
}
\seealso{
plotjchunk

simplebar

simplejbar
}

