% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/addons.R
\name{showtile}
\alias{showtile}
\title{Produces a heat map from 2 enumerations}
\usage{
showtile(enumdf, xlab, ylab, title = "", heatlow = "#FAFAFA",
  heathigh = "#2166AC")
}
\arguments{
\item{enumdf}{the data frame containing two enumerations}

\item{xlab}{the label for the x axis ("enum")}

\item{ylab}{the label for the y axis ("emum1")}

\item{title}{the title for the plot}

\item{heatlow}{the color value for the low end of values}

\item{heathigh}{the color value for the high end of values}
}
\description{
Given an enumeration data frame with "enum" and "enum1", it will return a ggplot
object of a heatmap comparing the two enumerations.
}

