% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/addons.R
\name{simplebar}
\alias{simplebar}
\alias{simplejbar}
\title{Plot a "simple" bar chart from an enumeration data.frame without details}
\usage{
simplebar(enumdf, title, ...)
}
\arguments{
\item{enumdf}{the data.frame containing the enumeration, expected to have the
text label in "enum" and "freq" with the frequecy variaable.}

\item{title}{the title for the plot}

\item{...}{optional variables passed into \code{ggplot::theme}}
}
\description{
This will return a ggplot render of an enumeration stored in a data.frame.
The enumeration could be from the basic \code{\link{getenum}} call, or
the \code{df} entry in the object returned by \code{\link{setjenum}}.
}
\details{
"simplebar" will return a plot that is simplified and polished.
It is just showing the percentage next to the
end of a solid bar for the value.

"simplejbar" is intended for an analyst to review as it showing counts and
percentages next to a solid bar with an error bar included.
}

