% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/addons.R
\name{slicetime}
\alias{slicetime}
\title{internal function, used in gettimeline.dbir}
\usage{
slicetime(timelabel)
}
\arguments{
\item{timelabel}{label to slice up}
}
\description{
internal function, used in gettimeline.dbir
}

