% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/csv.R
\name{sortvmatrix}
\alias{sortvmatrix}
\title{Internal: sort veris columns}
\usage{
sortvmatrix(x)
}
\arguments{
\item{x}{a matrix where columns are veris fields}
}
\description{
This will sort the columns of a matrix so Wade is happy.
}

