\name{summary.verisr}
\alias{summary.verisr}
\title{Internal Function.
This will create a mask of indexes based on filter
a list of matches.  This is intented to maintain the orginal
indexes of the veris object so further manipulation can be done.}
\usage{
  \method{summary}{verisr} (object, ...)
}
\arguments{
  \item{veris}{a verisr object}

  \item{enum}{the field to count}

  \item{value}{Displays a useful description of a ggplot
  object}

  \item{object}{veris object to summarise}

  \item{...}{other arguments ignored (for compatibility
  with generic)}
}
\description{
  Internal Function. This will create a mask of indexes
  based on filter a list of matches.  This is intented to
  maintain the orginal indexes of the veris object so
  further manipulation can be done.
}
\keyword{internal}

