% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{test_veris_ratio}
\alias{test_veris_ratio}
\title{Test the ratio of Ea to Eb}
\usage{
test_veris_ratio(
  chunk,
  Ea,
  Eb,
  ratio,
  direction,
  ci.level = 0.05,
  reps = 1000,
  quietly = FALSE,
  visualize = FALSE
)
}
\arguments{
\item{chunk}{getenumCI() output}

\item{Ea}{Enumeration A.  e.g. "action.Error"}

\item{Eb}{Enumeration B.  e.g. "action.Misuse"}

\item{ratio}{The ratio of Ea to Eb to test as a number. Must be greater than 1.}

\item{direction}{the direction to test ("greater" or "less")}

\item{ci.level}{the confidence level to test against}

\item{reps}{number of simulations to conduct}

\item{quietly}{do not produce textual output}

\item{visualize}{produce visual output}
}
\value{
a logical TRUE/FALSE to the hypothesis
}
\description{
Often we want to create quotes of the form "X is N times more/less likely than y in Z"
This function provides a test for such quotes.  It checks that the ratio is within the
ci.level credible interval and that the credible interval does not include 1
(indicating that we lack the ci.level confidence in the direciton of the ratio)
}
