% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{test_veris_time_stability}
\alias{test_veris_time_stability}
\title{Determine if the enumeration is changing over time}
\usage{
test_veris_time_stability(
  chunk,
  Ea,
  stable = NULL,
  direction = NULL,
  quietly = FALSE,
  visualize = FALSE
)
}
\arguments{
\item{chunk}{getenumCI() object}

\item{Ea}{Enumeration A.  e.g. "Error"}

\item{stable}{DEPRECIATED.  USE 'direction="equal"'. TRUE means "hypothesis is 'not changing'". FALSE means "hypothesis is 'changing'"}

\item{direction}{"greater" for increasing, "less" for decreasing, "equal" for "not changing"}

\item{quietly}{do not produce textual output}

\item{visualize}{produce visual output}
}
\value{
a logical TRUE/FALSE to the hypothesis
}
\description{
This is used by analyzing a dataset over time for an enumeration.
It functions by fitting a straight line to the values and 
checking if the confidence interval on the slope contains zero.
}
\details{
Technically instead of 'true/false', the language should really 
be along the lines of "we have evidence for the alternative  
hypothesis ..." or "we do not have evidence to go against our 
original null hypothesis ...", but for simplicity we have left
it the way it is.
}
\examples{
\dontrun{
tmp <- tempfile(fileext = ".dat")
download.file("https://github.com/vz-risk/VCDB/raw/master/data/verisr/vcdb.dat", tmp, quiet=TRUE)
load(tmp, verbose=TRUE)
chunk <- vcdb \%>\%
  filter(timeline.incident.year \%in\% 2015:2020) \%>\%
  filter(attribute.confidentiality.data_disclosure.Yes) \%>\%
  verisr::getenumCI2020("pattern.Lost and Stolen Assets", 
    by="plus.dbir_year", 
    ci.method="bootstrap", force= TRUE) \%>\%
  filter(!is.na(n))
  
  chunk
  
  verisr::test_veris_time_stability(chunk, "Lost and Stolen Assets", stable=TRUE)
}
}
