% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/csv.R
\name{veris2csv}
\alias{veris2csv}
\title{get matrix for CSV}
\usage{
veris2csv(dir = ".", veris = NULL)
}
\arguments{
\item{dir}{one or more directories with json files.}

\item{veris}{an optional veris object, if json already loaded.}
}
\description{
given a verisr object, this will create a matrix of the data
suitable for saving off as a CSV.
}

