% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verisr.R
\name{veriscol}
\alias{veriscol}
\title{Create a list of all column names expected from JSON schema.}
\usage{
veriscol(schema, a4)
}
\arguments{
\item{schema}{the merged veris schema in json}
}
\description{
Given a json schema for VERIS, this function will return a vector
of columns names to be set in the verisr object.  This is a wrapper 
around mkenums() which is a recursive function.  This will clean up
some of the one-off things we do after reading the schema.
}
\keyword{json}
