% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{flatten}
\alias{flatten}
\title{Convert a hierarchical verisr object to a flattened verisr object}
\usage{
flatten(veris, level = NULL, row.name = "extra.rowname")
}
\arguments{
\item{veris}{The verisr object to flatten}

\item{level}{Used during recursion}
}
\value{
flattened verisr object
}
\description{
(A flattened verisr object has no list columns)
}
