% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/matrix.R
\name{foldmatrix}
\alias{foldmatrix}
\title{Fold a verisr object into a numeric matrix of proportions}
\usage{
foldmatrix(veris, cols = NULL, rows = NULL, min = 1, clean = T)
}
\arguments{
\item{veris}{a veris matrix object}

\item{cols}{a vector of features to use in the columns, if left empty, it will use all the
logical columns in the input matrix.}

\item{rows}{a vector of column names or a named list of filters to use as the rows in the folded matrix}

\item{min}{the minimum number of matches a row must have to be considered, defaults to 1}

\item{clean}{if TRUE, this will look for anything that will mess up a veris or MDS function
and remove it, this includes empty rows and columns as well as columns with zero variance.}
}
\description{
given a verisr object and a list of features (columns) and objects (rows), this will
create a numeric matrix of the data suitable for further veris, hclust or other dist() analysis.
}
\details{
Each cell within the matrix will be a proprotion of incidents that either have the object
or match the pattern passed in a names list, depending if the rows argument is a vector of column
names or a names list of logical filters respectively.
}

