% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/addons.R
\name{getccc}
\alias{getccc}
\title{get the CCC values for each entry in a verisr object}
\usage{
getccc(veris)
}
\arguments{
\item{veris}{a verisr object}
}
\description{
Given a verisr object, calculate the scores for the complexity, completeness and context.
}
\details{
*Completeness* looks at the actor, action, asset and attributes, the core components
of the veris framework and produces a value between 0 and 1 for each of the categories.
Unknowns are treated as a zero and defined values are treated as a 1, the specific fields
in that categories are then averaged to produce the final value.

*Context* is computed the same way at completeness, but over the victim, timeline,
discvoery method and targeted data point.  Again this produces a value between 0 and 1.

*Complexity* is a simple count of fields that are defined and entered.  The theory
here being that lower scores may indicate that less overall data is known.
}

