% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getenum}
\alias{getenum}
\title{Get a vector of values from an enumeration}
\usage{
getenum(veris, enum, filter = NULL, add.n = T, add.freq = T)
}
\arguments{
\item{veris}{a verisr object}

\item{enum}{the field to count}

\item{filter}{limit what records are searched (optional)}

\item{add.n}{include a total count of variables found (denominator)}

\item{add.freq}{include a percentage (x/n)}
}
\description{
This will collect the values for an enumation
}
\details{
Note there are some special values that can be set as the enumeration
that are not obvious. :
* actor, action, attribute: will all return the next level down.  For example, just querying for "action" will return "malware", "hacking", and so on.
* action.variety: will return the variety enumerations across all actions (e.g. top N actions) (not in getenumby() yet)
* asset.assets: will return the type of assets, "Server", "Network, "User Dev" and so on
* victim.industry2: will return a short label of industries based on 2 values of NAICS code.
* victim.industry3: will return a short label of industries based on 3 values of NAICS code.
* pattern: will return the pattern the row is in.

Change: the "add.n" and "add.freq" options are now TRUE by default.
Change: the "primary" and "secondary" arguments were dropped.
}
\examples{
\dontrun{
hacking <- getenum(veris, "action.hacking.variety")
external <- getenum(veris, "actor.external.motive")
}
}

