% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getenumCI.R
\name{getenum}
\alias{getenum}
\title{Wrapper for getenumCI()}
\usage{
getenum(...)
}
\arguments{
\item{...}{getenumCI() parameters.  At a minimum, 'veris' (verisr object)
and enum (string)}
}
\description{
Wrapper for getenumCI()
}

