\name{getfilter}
\alias{getfilter}
\title{Internal Function.
This will create a mask of indexes based on filter
a list of matches.  This is intented to maintain the orginal
indexes of the veris object so further manipulation can be done.}
\usage{
  getfilter(veris, enum, value)
}
\arguments{
  \item{veris}{a verisr object}

  \item{enum}{the field to count}

  \item{value}{}
}
\description{
  Internal Function. This will create a mask of indexes
  based on filter a list of matches.  This is intented to
  maintain the orginal indexes of the veris object so
  further manipulation can be done.
}

