\name{getintenum}
\alias{getintenum}
\title{Internal Function.
This will iterate through the veris object and return
a list of matches.  This is intented to maintain the orginal
indexes of the veris object so further manipulation can be done.}
\usage{
  getintenum(veris, enum)
}
\arguments{
  \item{veris}{a verisr object}

  \item{enum}{the field to count}
}
\description{
  Internal Function. This will iterate through the veris
  object and return a list of matches.  This is intented to
  maintain the orginal indexes of the veris object so
  further manipulation can be done.
}

