% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/verisr.R
\name{getlast}
\alias{getlast}
\title{Get the last element from a column name}
\usage{
getlast(nm)
}
\arguments{
\item{nm}{the vector of column names}
}
\description{
Givn a vector with one or more column names (veris fields), this will
return the last string in the name, as it is seperated by [.].
}

