% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verisr.R
\name{getlogical}
\alias{getlogical}
\title{Convenience function to pull the logical columns from verisr object}
\usage{
getlogical(veris)
}
\arguments{
\item{veris}{the verisr object}
}
\description{
Given a verisr object this will return a vector of column names which are logical values.
}

