% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verisr.R
\name{getpattern}
\alias{getpattern}
\title{Determine the patterns from the 2014 Verizon DBIR}
\usage{
getpattern(sequence)
}
\arguments{
\item{sequence}{the unnested sequence column from a verisr object}

\item{rowname}{string identifying column uniquely identifying rows
Defaults to 'rowname'}
}
\description{
given a verisr object, this will determine which pattern the 
incident is in (or "Other" if no pattern is matched). Note the 
returned vector will be a factor with ordered levels for 
arranging the patterns in an order.
}
\examples{
data(veris.sample, package="verisr")

# produces a vector with 1-to-1 mapping to verisr object
pat <- getpattern(veris.sample)

# can summarize the results
table(pat)
}
