% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/addons.R
\name{industry2label}
\alias{industry2label}
\alias{industry2label.final}
\title{Convert 2 or 3 digit naics codes to matching industry title.}
\usage{
industry2label(industry, inc.code = T)
}
\arguments{
\item{industry}{a vector of industry codes, 2 or 3 digit}

\item{inc.code}{boolean if the codes should be included in label}
}
\description{
Given a vector of industry codes, these functions will return a matching
vector of the industry labels.
}
\details{
*industry2label* will not merge top level industries together (like manufacturing
or retail, which span 3 and 2 top level codes respectively).

*industry2label.final* will merge the top level categories.
}

