% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.verisr}
\alias{plot.verisr}
\title{Displays a four panel barplot of a verisr object}
\usage{
\method{plot}{verisr}(x, y, ...)
}
\arguments{
\item{object}{veris object to summarise}

\item{...}{other arguments ignored (for compatibility with generic)}
}
\description{
Displays a four panel barplot of a verisr object
}
\keyword{internal}

