% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verisr.R
\name{plot.verisr}
\alias{plot.verisr}
\title{Displays a four panel barplot of a verisr object}
\usage{
\method{plot}{verisr}(x, y, ...)
}
\arguments{
\item{...}{other arguments ignored (for compatibility with generic)}

\item{object}{veris object to summarise}
}
\description{
Displays a four panel barplot of a verisr object
}
\keyword{internal}
