% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verisr.R
\name{post.proc}
\alias{post.proc}
\title{Post process the veris object to add convenience fields.}
\usage{
post.proc(veris, veris_update_f)
}
\arguments{
\item{veris}{the verisr object}

\item{veris_update_f}{the pattern post-processing function}
}
\description{
Given a veris object this will populate several convenience fields
like the victim.industry2 and industry3,
}
\details{
Change in 1.1.3: now adds dummar vars for each pattern as "pattern.*"
}
