% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/graphics.R
\name{simplebar}
\alias{simplebar}
\title{Returns a simple bar plot of enumeration data.table}
\usage{
simplebar(enumdf, title = NULL, solidfill = "steelblue", ...)
}
\arguments{
\item{enumdf}{enumeration data table}

\item{title}{the title/label for the plot}

\item{solidfill}{a color to fill the bars with}

\item{...}{other arguments ignored (for compatibility with generic)}
}
\description{
given an enumeration data table and title, will plot a simple
bar chart, returning the ggplot object.
}

