% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summary.verisr}
\alias{summary.verisr}
\title{Displays a useful description of a verisr object}
\usage{
\method{summary}{verisr}(object, ...)
}
\arguments{
\item{object}{veris object to summarise}

\item{...}{other arguments ignored (for compatibility with generic)}
}
\description{
Displays a useful description of a verisr object
}
\keyword{internal}

