\name{summary.verisr}
\alias{summary.verisr}
\title{Displays a useful description of a ggplot object}
\usage{
  \method{summary}{verisr} (object, ...)
}
\arguments{
  \item{object}{veris object to summarise}

  \item{...}{other arguments ignored (for compatibility
  with generic)}
}
\description{
  Displays a useful description of a ggplot object
}
\keyword{internal}

