% Generated by roxygen2 (4.0.1): do not edit by hand
\name{veris2csv}
\alias{veris2csv}
\title{get matrix for CSV}
\usage{
veris2csv(dir = ".")
}
\arguments{
\item{dir}{one or more directories with json files.}
}
\description{
given a verisr object, this will create a matrix of the data
suitable for saving off as a CSV.
}

