% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{veris2csv}
\alias{veris2csv}
\title{get matrix for CSV}
\usage{
veris2csv(dir = ".")
}
\arguments{
\item{dir}{one or more directories with json files.}
}
\description{
given a verisr object, this will create a matrix of the data 
suitable for saving off as a CSV.
}
